[ci] Match signedness and type in test to avoid warnings in stm32.

+++ src/xpcc/container/test/bounded_deque_test.cpp
@@ -245,21 +245,21 @@ BoundedDequeTest::testOneElementQueue()
 {
 	xpcc::BoundedDeque<int16_t, 1> deque;
 
-	TEST_ASSERT_EQUALS(deque.getSize(), 0);
-	TEST_ASSERT_EQUALS(deque.getMaxSize(), 1);
+	TEST_ASSERT_EQUALS(deque.getSize(), 0U);
+	TEST_ASSERT_EQUALS(deque.getMaxSize(), 1U);
 
 	TEST_ASSERT_TRUE(deque.isEmpty());
 	TEST_ASSERT_FALSE(deque.isFull());
 
 	TEST_ASSERT_TRUE(deque.append(123));
-	TEST_ASSERT_EQUALS(deque.getSize(), 1);
+	TEST_ASSERT_EQUALS(deque.getSize(), 1U);
 
 	TEST_ASSERT_FALSE(deque.isEmpty());
 	TEST_ASSERT_TRUE(deque.isFull());
 
 	TEST_ASSERT_FALSE(deque.append(2));
 
-	TEST_ASSERT_EQUALS(deque.getFront(), 123);
+	TEST_ASSERT_EQUALS(deque.getFront(), int16_t(123));
 	TEST_ASSERT_EQUALS(deque.getBack(), deque.getFront());
 
 	deque.removeBack();
